devops/deployment/requirements.txt
flask==3.0.2
swagger-ui-py==23.9.23
gunicorn==21.2.0
docker==7.0.0
GitPython==3.1.42
